Rename user defined data reader file
