+++ dds/src/implementation/dds_impl/mod.rs
@@ -19,7 +19,7 @@ pub mod reader_factory;
 pub mod status_condition_impl;
 pub mod status_listener;
 pub mod topic_impl;
-pub mod user_defined_data_reader;
+pub mod user_defined_data_reader_impl;
 pub mod user_defined_data_writer;
 pub mod user_defined_data_writer_impl;
 pub mod user_defined_publisher;
