+++ dds/src/implementation/dds_impl/user_defined_subscriber.rs
@@ -17,7 +17,7 @@ use crate::{
 use super::{
     any_data_reader_listener::AnyDataReaderListener,
     domain_participant_impl::DomainParticipantImpl, topic_impl::TopicImpl,
-    user_defined_data_reader::UserDefinedDataReaderImpl,
+    user_defined_data_reader_impl::UserDefinedDataReaderImpl,
     user_defined_subscriber_impl::UserDefinedSubscriberImpl,
 };
